pyproject.toml
[project]
name = "pydantic-ai-provenance"
version = "0.1.0"
description = "Provenance tracking and citation verification for pydantic-ai agents"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Sumit Dugar" }]
requires-python = ">=3.12"
keywords = [
    "pydantic-ai",
    "provenance",
    "citation",
    "llm",
    "agents",
    "tracing",
    "attribution",
    "references",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic-ai-slim[anthropic,openai]>=1.80",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/sumitdugar/pydantic-ai-provenance"
Repository = "https://github.com/sumitdugar/pydantic-ai-provenance"
Issues = "https://github.com/sumitdugar/pydantic-ai-provenance/issues"
Changelog = "https://github.com/sumitdugar/pydantic-ai-provenance/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "pre-commit>=3.0",
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]

[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_ai_provenance"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["pydantic_ai_provenance"]

[tool.coverage.run]
source = ["pydantic_ai_provenance"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
